LogRender: only expand output table if emitting a path
When not emitting a path, `output.expand = True` causes log lines to be
filled with trailing whitespace to the width of the console.  This
commit means that we only expand the table when we are emitting path,
which is right-aligned.  Otherwise, we do not expand the table, avoiding
the trailing whitespace.

Fixes: #2647